Move tests to their own package
