+++ paperwork-integration/src/androidTest/java/hu/supercluster/paperwork/integration/test/ApplicationTest.java
@@ -1,4 +1,4 @@
-package hu.supercluster.paperwork.integration;
+package hu.supercluster.paperwork.integration.test;
 
 import android.app.Application;
 import android.test.ApplicationTestCase;
